#2: Delete finished#

+++ Database/Produkt.cs
@@ -4,6 +4,6 @@
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public int Preis { get; set; }
+        public decimal Preis { get; set; }
     }
 }
